backend/requirements.txt
asgiref==3.2.10
Django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
django-chartjs==2.2.1
